pyproject.toml
[project]
name = "comments-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"

dependencies = [
    "django",
    "djangorestframework",
    "psycopg2-binary",
    "pillow",
    "bleach>=6.3.0",
    "types-bleach>=6.3.0.20251115",
    "captcha>=0.7.1",
    "pytest>=9.0.2",
    "pytest-django>=4.12.0",
    "psycopg[binary]>=3.3.3",
    "django-extensions>=4.1",
    "graphviz>=0.21",
]

[dependency-groups]
dev = [
    "ruff",
    "black",
    "isort",
    "autoflake",
    "ty",
    "pre-commit",
    "mypy>=1.19.1",
    "django-stubs>=6.0.1",
]

[tool.mypy]
python_version = "3.14"
plugins = ["mypy_django_plugin.main"]

ignore_missing_imports = true
strict = false

[tool.django-stubs]
django_settings_module = "config.settings"
